# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases:
    get:
      operationId: getPurchases
      tags: [purchases]
      summary: Get all purchases
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: kategori
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of purchases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a new purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get a purchase by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePurchase
      tags: [purchases]
      summary: Update a purchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "200":
          description: Updated purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePurchase
      tags: [purchases]
      summary: Delete a purchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /purchases/bulk-delete:
    post:
      operationId: bulkDeletePurchases
      tags: [purchases]
      summary: Delete multiple purchases by IDs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRequest"
      responses:
        "200":
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /purchases/export/excel:
    get:
      operationId: exportPurchasesExcel
      tags: [purchases]
      summary: Export purchases as Excel file
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: kategori
          in: query
          required: false
          schema:
            type: string
        - name: ids
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated IDs for selective export
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /purchases/export/pdf:
    get:
      operationId: exportPurchasesPdf
      tags: [purchases]
      summary: Export purchases as PDF file
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: kategori
          in: query
          required: false
          schema:
            type: string
        - name: ids
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated IDs for selective export
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    BulkDeleteRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required:
        - ids
    Purchase:
      type: object
      properties:
        id:
          type: integer
        nomor:
          type: string
        tanggal:
          type: string
        keterangan:
          type: string
        jumlah:
          type: number
        satuan:
          type: string
        harga_satuan:
          type: number
        harga_total:
          type: number
        catatan:
          type: string
        kategori:
          type: string
        supplier:
          type: string
          nullable: true
        supplier_contact:
          type: string
          nullable: true
      required:
        - id
        - nomor
        - tanggal
        - keterangan
        - jumlah
        - satuan
        - harga_satuan
        - harga_total
        - catatan
        - kategori
    PurchaseInput:
      type: object
      properties:
        nomor:
          type: string
        tanggal:
          type: string
        keterangan:
          type: string
        jumlah:
          type: number
        satuan:
          type: string
        harga_satuan:
          type: number
        catatan:
          type: string
        kategori:
          type: string
        supplier:
          type: string
          nullable: true
        supplier_contact:
          type: string
          nullable: true
      required:
        - nomor
        - tanggal
        - keterangan
        - jumlah
        - satuan
        - harga_satuan
        - catatan
        - kategori
...