# lib/api-zod/src/generated/types/purchase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Purchase {
  id: number;
  nomor: string;
  tanggal: string;
  keterangan: string;
  jumlah: number;
  satuan: string;
  harga_satuan: number;
  harga_total: number;
  catatan: string;
  kategori: string;
  supplier?: string | null;
  supplier_contact?: string | null;
}